+++ mapentity/forms.py
@@ -30,12 +30,14 @@ class TranslatedModelForm(forms.ModelForm):
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
         # Track translated fields
+        self.orig_fields = list(self.fields.keys())
         self._translated = {}
-        self.replace_orig_fields()
-        self.populate_fields()
+
+        if 'modeltranslation' in settings.INSTALLED_APPS:
+            self.replace_orig_fields()
+            self.populate_fields()
 
     def replace_orig_fields(self):
-        self.orig_fields = list(self.fields.keys())
         # Expand i18n fields
         try:
             # Obtain model translation options
@@ -96,6 +98,7 @@ class MapEntityForm(TranslatedModelForm):
     fieldslayout = None
     geomfields = []
     leftpanel_scrollable = True
+    hidden_fields = []
 
     def __init__(self, *args, **kwargs):
         self.user = kwargs.pop('user', None)
@@ -223,8 +226,12 @@ class MapEntityForm(TranslatedModelForm):
                 field.fields = self.__replace_translatable_fields(field.fields)
                 newlayout.append(field)
             else:
+                # Add translated fields to layout
                 if field in self._translated:
-                    newlayout.append(self.__tabbed_layout_for_field(field))
+                    field_is_required = self.fields[f"{field}_{settings.MODELTRANSLATION_DEFAULT_LANGUAGE}"].required
+                    # Only if they are required or not hidden
+                    if field_is_required or field not in self.hidden_fields:
+                        newlayout.append(self.__tabbed_layout_for_field(field))
                 else:
                     newlayout.append(field)
         return newlayout